# This massively improves the performance of tight loops relying on a type() call.