--- a/include/simdjson/document.h
+++ b/include/simdjson/document.h
@@ -44,7 +44,7 @@ constexpr const uint64_t JSON_VALUE_MASK = 0x00FFFFFFFFFFFFFF;
 constexpr const uint32_t JSON_COUNT_MASK = 0xFFFFFF;
 
 /**
- * The possible types in the tape. Internal only.
+ * The possible types in the tape.
  */
 enum class tape_type {
   ROOT = 'r',
@@ -93,15 +93,17 @@ namespace simdjson::dom {
  * This is the type it is most easily cast to with get<>.
  */
 enum class element_type {
-  ARRAY,     ///< dom::array
-  OBJECT,    ///< dom::object
-  INT64,     ///< int64_t
-  UINT64,    ///< uint64_t: any integer that fits in uint64_t but *not* int64_t
-  DOUBLE,    ///< double: Any number with a "." or "e" that fits in double.
-  STRING,    ///< std::string_view
-  BOOL,      ///< bool
-  NULL_VALUE ///< null
+  ARRAY = 7,     ///< dom::array
+  OBJECT = 11,    ///< dom::object
+  INT64 = 10,     ///< int64_t
+  UINT64 = 5,    ///< uint64_t: any integer that fits in uint64_t but *not* int64_t
+  DOUBLE = 2,    ///< double: Any number with a "." or "e" that fits in double.
+  STRING = 6,    ///< std::string_view
+  BOOL = 4,      ///< bool
+  NULL_VALUE = 12 ///< null
 };
+
+/* We choose the integers 7, 11, and so forth in element_type so that the matching tape values (", {,..,) are mapped to them using the x -> x % 14 function. */
 
 /**
  * JSON array.