Before:
```
numbers_type_scan                       83977 ns        83978 ns         8325
numbers_type_size_scan                  78568 ns        78570 ns         8910
```

After:

```
numbers_type_scan                       31775 ns        31776 ns        21990
numbers_type_size_scan                  33834 ns        33835 ns        20686
```

The main "trick" is to avoid relying on a switch-case when converting the type read
from the tape (an ascii character value) to a C++ enum class. Instead we do some
branchless arithmetic.

Again, I lack any kind of good metric, so I am going out of instinct and experience. We need instrumented benchmarks where I can see what is going on. Anyhow, for now there are still low-hanging fruits, apparently.