@lemire 2020-04-15
It is still a bit inconvenient the way we need to go from one enum to the other. The only difference really is that @jkeiser introduced the BOOL type which makes sense conceptually but forced me to do some acrobatics. So instead of being 2-3 instructions, a call to type() becomes maybe 6 to 10 instructions... but at least it is branchless. 

There might be a cleverer way...