@lemire 2020-04-15
One easy way to fix this mess is to change the tape to better match the API.