@lemire 2020-04-15
The cmov approach is at least as fast and much simpler. Let us go with that.