@lemire 2020-04-16
This is clean and much faster. Let us merge.